Install latest version of pip in pyvenv
The new version of the cryptography package (3.4.3) has a build-time
dependency on Rust because the maintainers rewrote some internals.
Because Rust is not installed in nso-docker building fails for obvious
reasons. However, the maintainers also provide wheel packages! Older
versions of pip (currently 18.x) does not download the wheel package
correctly. This is apparently fixed in newer pip versions.