Improve IPv6 networking
This fixes a number of issues around IPv6 networking:
- simplify address variable use & computation
- remove global IPv6 connectivity in most cases
- fix OS X compatibility

Simplify the computation of the IPv6 address. The reason to why I chose to
separately compute the IPV6_NET first was that we used it twice, in providing
the arguments to:

  docker network create --ipv6 --subnet $(IPV6_NET):/64 --gateway $(IPV6_NET):1

Defining IPV6_NET as a normal variable in Make means it is lazily evaluated, so
it becomes evaluated when needed, i.e. when expanding that statement. Initially
I believed that would work out of the box, because it would make sense for Make
to cache the value, at least if used multiple times in the same statement.
That's not true however, the computed value is not cached and thus, we compute
it twice for the above expression. Since it is random, we end up with two
different random values.

The variable assignment in make can be made non-lazy by doing:

  IPV6_NET:= ...

However, that means we would always compute the address, even when we are
running some completely different make target - not good as it will slow down
the execution of the makefile. Probably not noticeable for a single variable
like this, but it's an anti-pattern and down the line, with many such variables,
it would become noticeable.

This is now addressed. We don't explicitly specify the gateway, so the variable
is only needed once. Further, we later need the IPv6 network, but at that point,
as the network is created, we can dig it out of docker instead. The computation
now happens once inline in bash.

It was reported that the IPv6 part didn't work on OS X. This is because tr
assumes UTF-8 on OS X (depends on locale settings but most users typically use
UTF-8) and when reading /dev/random we get bits we don't recognize. By setting
LC_CTYPE (character types) to C we allow all byte values.

We want IPv6 enabled per default to enable local IPv6 connectivity. We use ULA
addressing, since we want unique addresses (in case of multiple parallel
testenvs). This means we don't have a valid globally routable IPv6 address and
cannot try to use them to communicate with the Internet. However, Docker sets up
routing per default, so containers will attempt to use IPv6, which will fail.
The result is that most operations take a long time before timing out. We now
avoid this situation by removing the IPv6 default route when the addressed used
are not in the public unicast space (2000::/3).