Avoid ncs-backup failure
ncs-backup in later versions of NSO (5.5.1 and possibly other branches)
will fail if the packages directory is missing from the run-dir, i.e.
/nso/run/packages.

From the very beginning of NSO in Docker, we have encouraged packages to
be placed in the container image, in /var/opt/ncs/packages, whereas the
persistent state, like CDB and netconf notification replay logs, are in
the normal run-dir, which is mounted in from a volume for persistence.
Thus we haven't made use of the packages directory that normally exists
in the run-dir.

Further, the run-dir is normally created at NSO installation time and
will then contain all the usual directories. Since we encourage the /nso
directory to be mounted from a volume, we don't assume that it holds an
existing run-dir. Rather, we create one if it doesn't exist. Since we
haven't needed a package directory, we haven't created it.

However, since some time back, we also support loading packages from the
run-dir, so if someone wants to place packages there, that is possible.
It is bad for all sorts of different reasons, but it could provide for a
simpler migration path. Thus we load packages both from
/nso/run/packages and from /var/opt/ncs/packages.

Creating /nso/run/packages if it doesn't exist is natural since it makes
it more visible that users can place their packages there... and it has
the benefit of conforming to what ncs-backup in NSO 5.5.1 expects.

We always create these directories if they do not exist, so for anyone
that has started using NSO in Docker with NSO 5.5.1 the fix should
merely me to rebuild their images with the latest version of nso-docker.

Fixes #214.