Prefer CI_DEFAULT_BRANCH to static 'master'
This builds on the previous commit, changes additional places and above all in
the skeletons. Some wordings in READMEs etc have been fixed to mention both main
and master. I don't just want to remove the mentioning of 'master' because many
people will still be used to it and since the purpose is to give a sense of
reference we can't just pretend the concept of a master branch never existed in
git.

There are still occurrences of master, which mostly is due to the default branch
on this repo still being master. That needs to be addressed separately.