Stronger wording for CI syntax
I don't want to offer this "structured" notation as yet another alternative,
rather it should be the go to option and the sort of default - I mean, I think
it is normal to mirror the nso-docker repo on to your environment, so let the
wording reflect that.

I've also removed the example of using a URL for your own nso-docker repo. I
think people are clever enough to understand what they need to do if they want
URL style...