Reduce startup time by speeding up CDB debug-dump
I have a vague memory of using | head to terminate the cdb debug dump
early on large CDB files but it's apparently not a trick I've applied
here. Starting up on a 1GB CDB would take like an hour on my old
workstation simply because of how poorly this was implemented.

With this fix, we drastically reduce the time to inspect CDB to a second
or so.