Add test suite embryo for skeletons
This adds a framework for testing our skeletons. The initial test is a build &
test of a simple python package. I've also added a no-src test which tests
building a package that doesn't have a src directory. That currently fails and
thus the test is currently commented out.

For the NED skeleton we are using a separate NED package since it otherwise
fails compiling the netsim etc.