Fix docker build cache disabling
In our Dockerfile we have a common 'build' stage whose output is reused in the
other stages, like 'package', 'testnso' or similar. Having a common build stage
means we reduce compilation, thus speeding up the build and also avoid
inconsistency issues, like compiler bugs when compiling multiple times.

In CI, we want to specifically not cache builds so as to install new
dependencies or recompile things with later tools, thus when it is detected we
run in CI (through the environment variable CI being set), we set
DOCKER_BUILD_CACHE_ARG=--no-cache. Due to the way how it was applied in the
build make target, we would actually rebuild each target, and since each target
include things from the 'build' target, the 'build' target would be rebuilt
multiple times. For lengthy compilations or installations of dependencies from
the Internet, this could take a considerable amount of time.

Besides it being slow to build everything multiple times, rebuilding also opens
up for inconsistencies between the produced docker images as it is imaginable an
upstream dependency changes between building say the 'testnso' and 'package'
image. Our incorrect application of the --no-cache argument meant we failed to
deliver on the consistency we set out to achieve.

Now we instead apply --no-cache for just the 'build' target. This is effectively
invalidating the "cache" for the other build targets like 'testnso' or
'package', since if the output of the 'build' target changes, then the 'testnso'
and 'package' images will in fact be rebuilt.

I don't see a need to specifically skipping the cache for just the 'testnso',
'package' target or similar.