NED skel: fix add-device on NSO 4.x
We have to strip away the ned-id, which we did do in the old code but had an
incorrect assumption on that the device must be a CLI device. It's not fixed to
support any device-type. The lack of the ned-id attribute means we have to
explicitly list them, but that's sort of okay. No new types have been added in
forever.