Test persistence of crypto keys
This adds a test to ensure we can decrypt encrypted values in a multi-version
environment, which essentially boils down to ensuring that we use the same
crypto keys, that is, they are persisted to disk. That's currently not the case,
so this test is expected to fail.

To get the encrypted value in cleartext, it has to be decrypted. There is no
decryption function available in the CLI, so we implement one in Python. The
standard testpkg doesn't use Java or Python, so I switched over to using the
testpkg-python package instead and modified it to be able to store our values
etc.