Persist ncs.crypto_keys
The encryption keys for NSO are stored in a separate file, which we have, rather
embarrassingly, failed to persist. It is now persisted!

Just like for SSH keys and SSL certs, we check for its existence on startup and
generate it, if it does not exist.

On startup, we transfer the information in /nso/etc/ncs.crypto_keys to the NSO
configuration. This is done differently depending on the version of NSO. NSO 5.x
uses /etc/ncs/ncs.crypto_keys whereas NSO 4.x has the encryption keys inlined
into ncs.conf. NSO 5.3 and newer supports AES256 whereas older NSO 5.x version
will complain about the presence of the ASE256CFB128_KEY in the ncs.crypto_keys,
thus we have to strip it out.

We also remove /etc/ncs/ncs.crypto_keys from the base image since it could cause
confusion.