Add test case for new config mgmt approach
The idea is to allow injection of a ncs.conf by placing it on the persistent
volume mounted at /nso. We test this by first starting up a container and
enabling the HTTP interface. The container has a persistent volume mounted at
/nso and we copy over the generated config file from /etc/ncs/ncs.conf to
/nso/etc/ncs.conf and then shut down the container. Next we start a new
container, this time without enabling the HTTP interface but it has the same
volume mounted. It should detect the persisted /nso/etc/ncs.conf and use it!
We repeat the same procedure and start a third container, but this time we
enable config mangling and explicitly disable the HTTP interface.