Fix pre/post start invocation
Not sure why we sourced these scripts in the past. It makes it impossible to
implement an early exit, since by sourcing the script is read in the current
environment and thus we would exit from the current environment. What we want is
to just execute the scripts like any other binary.