Switch to using bash
We have bash in the container image so might as well use it. We actually relied
on bashism in this file and as we were sourced from the run-nso script, we
inherited the bash interpreter that it uses. When that was changed to run this
script, we instead ended up with sh and things stopped working.