Always set config & avoid bashism
After the change to run pre/post scripts rather than sourcing them, thus
changing the interpreted to sh instead of bash, we would fail as these if
conditions were implemented in bash forgiving way (should be = and not ==).

Anyway, we also want to always be able to run these modification commands as
with the new approach to config mangling, it might be that a persisted ncs.conf
has HTTP enabled and we want to explicitly disable it.