Add test case for mounted configuration
Similar to having ncs.conf on a persistent volume, this will test the capability
of directly mounting in ncs.conf to /etc/ncs/ncs.conf. It's another option for
convenience (since setting up a volume and injecting a ncs.conf requires
multiple commands etc).

The overall process is quite like when testing on a volume, the difference is
that we instead copy the config and then mount it in directly to
/etc/ncs/ncs.conf.