Support mounting in ncs.conf
This now adds a third way of dealing with ncs.conf: it is now possible to
directly mount a ncs.conf to /etc/ncs/ncs.conf in the container.

We achieve this by renaming the factory installed ncs.conf to
/etc/ncs/ncs.conf.in, thus if we start up with a /etc/ncs/ncs.conf we know it
must be mounted and can act accordingly. If /etc/ncs/ncs.conf does not exist, we
will proceed with the normal method of generating a ncs.conf, but based on
/etc/ncs/ncs.conf.in

In order of preference
- directly mounted /etc/ncs/ncs.conf (no mangling per default)
- /nso/etc/ncs.conf on a volume (no mangling per default)
- /etc/ncs/ncs.conf(.in) from installed NSO + mangling