Make config mangling idempotent
So that was surprisingly simple - I've balked for a long time, thinking
this would require a lot more magic but it turns out to that a) a lot of
patches are already idempotent - just setting a value and b) making the
more advanced ones idempotent is a matter of XPath hackery, which is
quite simple.

The core lies in selecting nodes for edits by excluding ones with
children that we were going to insert.