Map NSO_VERSION=DOCKER_TAG for skel tests
It's actually slightly confusing how NSO_VERSION refers to the NSO
version when building this repo, like 5.5.1 is the 5.5.1 installer...
but we produce something that looks like 5.5.1-1234 (CI pipeline ID)...
then in skeletons we use the same NSO_VERSION variable, but here we use
it to refer to the nso-docker container images, which might or might not
have a suffix.

Anyway, for testing the skeletons we want to use the version of
nso-docker that we've just built, thus mapping NSO_VERSION=DOCKER_TAG!

Not sure if this is the right place to do that mapping though, but
should work for now I think.