Improve dead NSO detection during startup
Once we've started NSO we wait for it to start before running the post-start
scripts. To avoid just sleeping for a fixed period of time, we run ncs
--wait-started which will instead attempt to communicate with NSO and wait for
NSO to report that it has started up.

The problem here is that with ncs --wait-started 600 we wait up to 600 seconds.
If NSO has died, for example by having a corrupt CDB, then there's no chance of
NSO recovering and there's no point in waiting for that long. Thus, we now have
a loop where we check that the NSO PID is alive and then ask if NSO has started.

Since we test the NSO exit code when started with a corrupt CDB, our own CI
tests have actually been plagued by the extra time it took for the container to
realize NSO is dead, and thus our tests have taken quite a long time. This
should speed it up by roughly 600 seconds!

Also convert the test to directly use the cisco-nso-base image rather than the
test image.