Add ability to restart NSO process
Sending SIGHUP to the container, like with docker kill --signal=HUP ..., will
now be handled by the init script in the container and restart NSO.

In a pure containeresque environment, this is generally frowned upon. We should
be restarting the container and not a process in the container. However, we have
already implemented certain measures that are similarly non-containeresque and
by allowing our users to opt-in to some, and not all, idiomatic container
behavior on day 1, we enable a gradual transition from where users are today to
a container native environment. That's a good thing overall and in the general
direction of where we want to be.