Remove wait-healthy from CI
I don't think this target belongs being called directly from CI. We don't want
to expand the interface between CI and the make targets. If anything we want to
reduce it. Calling testenv-wait-healthy seems superfluous. Either the start call
can wait till it is done or probably the better option is to let testenv-test
wait for it to complete...