Add support for building & packaging multiple NED versions
In NSO 5.x (with CDM support) changing the NED for a device requires using the
built-in migrate action. At the time of the migration both old and new versions
of the NED must be loaded. This workflow is now enabled by the NID/NED skeleton.
It is possible to build and package multiple NED versions by just including
them in the packages/ directory. Some notes on the resulting docker images:
- "repo/package" contains all versions (all ned-ids)
- "repo/package-ned-id-1.23" contains just the package with ned-id ned-id-1.23
- "repo/testnso" contains all versions
- "repo/netsim" contains just the latest version
- "repo/netsim-ned-id-1.23" contains just the package with ned-id ned-id-1.23