Add env var for controlling HA
This adds the environment variable HA_ENABLE which, as the name
suggests, controls whether HA is enabled or not in ncs.conf. As other
options, it is implemented idempotently and thus will not add duplicate
ha entries to the config. It will also always set the value, so that it
does not rely on the default config being either true or false.