Guide user based on NSO_VERSION
I've shoulder surfed people that set NSO_VERSION to build an NSO In
Docker image for a specific version and then proceed to run 'make',
which runs the default target 'all', which in turn builds images for all
NSO versions we can find in the NSO_INSTALL_FILES_DIR.

We now error out if a user tries to run the 'all' target with
NSO_VERSION set. It is expected that if NSO_VERSION is set, the user
intends to build for a single version. If they indeed do want to build
for all versions, simply unset NSO_VERSION first.

This is backwards incompatible but I think it's fine.