Rewrite test-backup-restore
The test is rewritten such that it does not use a persistent volume for
the first part of the test. We write some data to CDB, take a backup and
copy the file to a temporary location outside of the container.
The restore in the second part includes two volumes - one mounted to
/nso for the run directory and the second mounted to /etc/ncs for the
config directory. The restore command is executed in a helper container.
When we start the NSO container with the restored data the run volume
mount remains the same (/nso) while the config volume is mounted to
/nso/etc.