Do not unconditionally run add --platform
The --platform argument is relatively new and will cause a problem when
trying to run on older machines.

The change here is simple. We check if we are running on anything other
than x86_64, in which case we assume the --platform argument exists.
This should pretty much only be valid when running on an Apple M1 or
some ARM server and if Docker exists on such a box, it is
likely (always?) of a new enough version that it supports --platform.

I had a feeling this was going to pop up as a problem but somehow
figured that "eh, no one runs *that* old docker versions... right?". I
was wrong. I have no statistics of course but tip of iceberg and all
that. Unfortunately I'm unable to test that this actually works on an M1
as I don't have access to one right now but this really should work.