Improve use of --platform & DOCKER_BUILD_ARGS
Condition the use of --platform to when we are not running on x86_64.
NSO is only compiled for x86_64, so if we are on some other
architecture, we ask docker to emulate x86_64 for us. Magic! We had this
statically added before but we got some feedback that people are using
old enough Docker versions that they do not have the --platform arg and
thus run into errors. It is now conditioned on the platform. I think any
version of Docker that is running on something that is not x86_64 will
have support for the --platform argument.

Move DOCKER_BUILD_ARGS to nidcommon, since it is common across all
skeletons. It now includes the DOCKER_PLATFORM_ARG and correctly uses +=
to add things. We previously used := twice, causing us to miss the
platform arg.

Building of the netsims now also uses DOCKER_BUILD_ARGS, as it should!

Fixes #233.