Rebuild pyvenv(-dev) in testenv rebuild
This enables us to modify the requirements(-dev)?.txt files and install new
packages in a running testenv. Previously one had to rebuild the docker image
and restart the testenv.
The development venv (pyvenv-dev) is not included in the image at all. If
requirements-dev.txt is present the venv is now created in the testenv container
the first time rebuild runs.