Stop startup if SSH keys / SSL certs are present in old location
The files were moved to subdirectories in /nso/etc (NCS_CONFIG_DIR). We don't
want to silently copy or move the files to the new location because that could
lead to surprises down the road when the user wants to change the secrets. It is
better to inform the users of the new location and instruct them to move the
files themselves.