Re-mangle startup config in backup & restore test
The backup file includes the contents of NCS_CONFIG_DIR, including the mangled
ncs.conf. When we restore the backup the mangled file is placed back in the
config dir (and then /nso/etc). Normally reusing the mangled config is fine and
actually the way one would restore a backup anyway. But when restoring the
backup to a newer version of NSO there may be forward incompatibilities the
script does not know how to resolve. Let's just remove the config in that case!