Ensure we are passed the linux installer
We've seen users attempt to build the NSO container images based on the
Darwin / Mac installer, which is a perfectly natural thing to attempt,
if you are running Docker on Mac. It is not obvious that Docker is
actually running a Linux VM under the hood and that you therefore must
use the Linux installer.

We guide the user by throwing an error if anything that is not the Linux
installer is passed!