Add CI job retries
We only retry jobs if they fail due to some form of infrastructural
failure, like the runner failed to start up. If the actual test failed,
like some part of the script returned an error, we don't retry that. We
don't want to encourage users to write flaky tests and rely on the
retries, this is just for covering annoying intermittent problems with
build infrastructure.