Filter containers in wait-healthy on label
Rather than filtering on the name and then relying on the CNT_PREFIX
being used as a prefix, we use a label, which should be slightly more
robust.

vrnetlab isn't mentioned anymore!

There was a hard-coded check for $(CNT_PREFIX)-nso, i.e. the sort of
default name for an NSO instance. This has been removed. For a testenv
with multiple NSOs it might be natural to name all of them with unique
names, like nso-primary, nso-secondary or similar. Thus, such a check
with a hard-coded name makes little sense.

We already have a check that looks at all containers, this will now also
be used to check the NSO containers, making them less of a special case.