Optimize testenv-rebuild initial rebuild
When running testenv-rebuild for the first time in a testenv it is
likely the list of rsync'd files will include files where the contents
has not really changed but the files have different (newer) mtimes.
If these false positives include *.yang files the build script will
execute a full packages reload at the end. This is a time-consuming
operation when many packages are loaded and is unnecessary.

To avoid this we run the first rsync with checksumming instead, then
after it is done and for the rest of the container lifetime we run it in
default (mtime + size comparison) mode.